requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dill==0.3.9
docplex==2.28.240
fastdtw==0.3.4
Flask==3.0.3
frozendict==2.4.6
graphviz==0.20.3
h5py==3.12.1
html5lib==1.1
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.57.2
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lark==1.2.2
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mqt.bench==1.1.7
multitasking==0.0.11
Nasdaq-Data-Link==1.0.4
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pbr==6.1.0
peewee==3.17.7
platformdirs==4.3.6
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyscf==2.7.0
pyspnego==0.11.1
python-dateutil==2.9.0.post0
pytket==1.33.1
pytket-qiskit==0.58.0
pytz==2024.2
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-algorithms==0.3.1
qiskit-finance==0.4.1
qiskit-ibm-runtime==0.31.0
qiskit-nature==0.7.2
qiskit-optimization==0.6.1
qwasm==1.0.1
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
scipy==1.14.1
six==1.16.0
soupsieve==2.6
stevedore==5.3.0
symengine==0.13.0
sympy==1.13.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
yfinance==0.2.46
